==> prediction/1-prepare.ipynb <==
# Prepare the selected metapaths for feature computation for all compound-disease pairs

import json
import itertools
import bz2
import configparser

import pandas
from statsmodels.sandbox.stats.multicomp import multipletests

## Read node info

# Read compound and disease degrees
compound_df = pandas.read_table('../summary/compounds.tsv')
disease_df = pandas.read_table('../summary/diseases.tsv')
len(compound_df), len(disease_df)

indication_df = pandas.read_table('../summary/indications.tsv')
indication_df['category'] = 'DM'
indication_df = indication_df[['compound_id', 'disease_id', 'category']]
indication_df.head(2)

reach_df = pandas.read_csv('../all-features/data/reach.csv', index_col=0).reset_index(drop=True)
reach_df.head(2)

rows = list()
for c, d in itertools.product(compound_df.itertuples(), disease_df.itertuples()):
    row = c.compound_id, c.compound_name, d.disease_id, d.disease_name
    rows.append(row)
pair_df = pandas.DataFrame(rows, columns=['compound_id', 'compound_name', 'disease_id', 'disease_name'])
pair_df = pair_df.merge(indication_df, how='left')
pair_df['status'] = (pair_df.category == 'DM').astype(int)
pair_df.head(2)

pair_df['status'].value_counts()

len(pair_df)

with bz2.open('features/compound-disease-pairs.tsv.bz2', 'wt') as write_file:
    pair_df.to_csv(write_file, sep='\t', index=False)

## Select metapaths

auroc_df = pandas.read_table('../all-features/data/feature-performance/auroc.tsv')
reject, pvals_corrected, alphacSidak, alphacBonf = multipletests(auroc_df.pval_delta_auroc, method='fdr_bh')
auroc_df['fdr_delta_auroc'] = pvals_corrected
auroc_df.head(2)

# standard is 0.55
whitelist_df = auroc_df.query(
    "rdwpc_auroc > 0.54"
    " and delta_auroc > 0"
    " and fdr_delta_auroc < 0.05"
    " and pdwpc_primary_auroc > 0.5"
).copy()
whitelist_df['feature'] = 'dwpc_' + whitelist_df['metapath']
whitelist_df.head(2)

metapaths = set(whitelist_df.metapath)
len(metapaths)

metapaths

## Metaedges in chosen metapaths

m2m_df = pandas.read_table('../all-features/data/metaedge-in-metapath.tsv')
m2m_df = m2m_df.query("metapath in @metapaths")

# Counting multiple occurrences in the same metapath
m2m_df.metaedge.value_counts().reset_index()

# Counting only one metaedge occurrence per metapath
m2m_df.drop_duplicates().metaedge.value_counts().reset_index()

# Number of included metaedges
m2m_df.metaedge.nunique()

with open('../all-features/data/metapaths.json') as read_file:
    metapath_obj = json.load(read_file)
metapath_obj = [x for x in metapath_obj if x['abbreviation'] in metapaths]
with open('features/metapaths.json', 'wt') as write_file:
    metapath_obj = json.dump(metapath_obj, write_file, indent=2, sort_keys=True)

total_queries = len(metapaths) * len(pair_df)
print('{:,} total queries'.format(total_queries))
