==> prediction/2-extract.ipynb <==
# Extract DWPCs for a select set of metapaths but for all compound-disease pairs

import json
import threading
import concurrent.futures
import bz2
import csv
import time

import pandas
import py2neo

## Connect to Neo4j

# Override the default py2neo timeout
py2neo.packages.httpstream.http.socket_timeout = 1e8

with open('../all-features/servers.json') as read_file:
    instances = json.load(read_file)

for instance in instances:
    if instance['name'] == 'wikidata-v0.1':
        uri = 'http://localhost:{}/db/data/'.format(instance['port'])
        bolt_port = 7690 + (instance['port'] % 10)
        neo = py2neo.database.Graph(uri, bolt = True, bolt_port = bolt_port)
        break
neo

## Read metapaths

with open('features/metapaths.json') as read_file:
    metapaths = json.load(read_file)

metapaths.sort(key=lambda x: x['join_complexities'][0])
len(metapaths)

pair_df = pandas.read_table('features/compound-disease-pairs.tsv.bz2')
pair_df.tail(2)

reach_df = pandas.read_csv('../all-features/data/reach.csv', index_col=0).reset_index(drop=True)
print(len(reach_df))
reach_df.head(2)

# Total number of queries
total_queries = len(metapaths) * len(reach_df)
print('{:,} total queries'.format(total_queries))

## Query setup

pairs = list(reach_df.itertuples())

def generate_parameters(max_elems=None):
    """Generate compound, disease, metapath combinations"""
    n = 0
    for metapath_dict in metapaths:
        metapath = metapath_dict['abbreviation']
        query = metapath_dict['dwpc_query']
        for pair_info in pairs:
            if max_elems is not None and n == max_elems:
                break
            yield {
                'neo': neo,
                'hetnet': 'hetio-ind',
                'compound_id': pair_info.compound_id,
                'disease_id': pair_info.disease_id,
                'metapath': metapath,
                'query': query,
                'w': 0.4,
            }
            n += 1

def compute_dwpc(neo, hetnet, query, metapath, compound_id, disease_id, w):
    """Execute the neo4j query and write results to file"""
    start = time.time()
    results = neo.data(query, source=compound_id, target=disease_id, w=w)
    record = results[0]
    seconds = '{0:.4g}'.format(time.time() - start)
    row = hetnet, compound_id, disease_id, metapath, record['PC'], w, '{0:.6g}'.format(record['DWPC']), seconds
    with writer_lock:
        writer.writerow(row)

## Execute queries

%%time

# Parameters
workers = 32
max_elems = None

# Prepare writer
path = 'features/dwpc.tsv.bz2'
write_file = bz2.open(path, 'wt')
writer = csv.writer(write_file, delimiter='\t')
writer.writerow(['hetnet', 'compound_id', 'disease_id', 'metapath', 'PC', 'w', 'DWPC', 'seconds'])

# Create ThreadPoolExecutor
executor = concurrent.futures.ThreadPoolExecutor(max_workers=workers)
writer_lock = threading.Lock()

# Submit jobs
n_queries = 0
for params in generate_parameters(max_elems):
    while executor._work_queue.qsize() > 10000:
        print('Submitted queries: {} ({:.4%})'.format(n_queries, n_queries / total_queries), end='\r')
        time.sleep(1)
    executor.submit(compute_dwpc, **params)
    n_queries += 1

# Shutdown and close
executor.shutdown()
write_file.close()

n_queries
